tagsearch moved to header
